bot/requirements.txt
python-dotenv==0.19.2
aiogram==2.19
psycopg2-binary==2.9.3
openpyxl==3.0.9
pandas==1.4.2
pytz==2022.1
pygsheets==2.0.5
loguru